# hellokratos/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /rider/v1/code/send:
        post:
            tags:
                - Auth
            description: 发送验证码
            operationId: Auth_SendCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.SendCodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.SendCodeReply'
...
    /rider/v1/login/password:
        post:
            tags:
                - Auth
            description: 密码登录
            operationId: Auth_LoginByPassword
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.LoginByPasswordRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.LoginReply'
    /rider/v1/login/phone:
        post:
            tags:
                - Auth
            description: 手机号+验证码登录
            operationId: Auth_LoginByPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.LoginByPhoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.LoginReply'
    /rider/v1/login/third:
        post:
            tags:
                - Auth
            description: 第三方登录
            operationId: Auth_LoginByThirdParty
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.LoginByThirdPartyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.LoginReply'
    /rider/v1/logout:
        post:
            tags:
                - Auth
            description: 登出
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.LogoutReply'
...
    /rider/v1/register:
        post:
            tags:
                - Auth
            description: 注册
            operationId: Auth_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.LoginReply'
    /rider/v1/token/refresh:
        post:
            tags:
                - Auth
            description: 刷新token
            operationId: Auth_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.RefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.LoginReply'
...
components:
    schemas:
...
        api.rider.v1.LoginByPasswordRequest:
            type: object
            properties:
                phone:
                    type: string
                password:
                    type: string
            description: 密码登录请求
        api.rider.v1.LoginByPhoneRequest:
            type: object
            properties:
                phone:
                    type: string
                code:
                    type: string
            description: 手机号登录请求
        api.rider.v1.LoginByThirdPartyRequest:
            type: object
            properties:
                platform:
                    type: string
                code:
                    type: string
                state:
                    type: string
            description: 第三方登录
        api.rider.v1.LoginReply:
            type: object
            properties:
                userId:
                    type: string
                token:
                    type: string
                refreshToken:
                    type: string
                expiresIn:
                    type: string
                userInfo:
                    $ref: '#/components/schemas/api.rider.v1.UserInfo'
            description: 登录响应
        api.rider.v1.LogoutReply:
            type: object
            properties:
                success:
                    type: boolean
            description: 登出响应
        api.rider.v1.LogoutRequest:
            type: object
            properties:
                userId:
                    type: string
            description: 登出请求
...
        api.rider.v1.RefreshTokenRequest:
            type: object
            properties:
                refreshToken:
                    type: string
            description: 刷新token请求
        api.rider.v1.RegisterRequest:
            type: object
            properties:
                phone:
                    type: string
                password:
                    type: string
                code:
                    type: string
                nickname:
                    type: string
            description: 注册请求
        api.rider.v1.SendCodeReply:
            type: object
            properties:
                requestId:
                    type: string
                expireSeconds:
                    type: integer
                    format: int32
            description: 发送短信响应
        api.rider.v1.SendCodeRequest:
            type: object
            properties:
                phone:
                    type: string
                type:
                    type: string
            description: 短信发送请求
...
        api.rider.v1.UserInfo:
            type: object
            properties:
                userId:
                    type: string
                phone:
                    type: string
                nickname:
                    type: string
                avatar:
                    type: string
                status:
                    type: integer
                    format: int32
            description: 用户信息
...